requirements.txt
#python==3.11

streamlit==1.31.1
SpeechRecognition==3.10.1
edge-tts==6.1.9
pygame==2.5.2
groq==0.4.2
python-dotenv==1.0.1
PyAudio==0.2.14